Expanded migration config validations
Migration configuration both in the KubeVirt CR and MigrationPolicy now has expanded validations for its fields to ensure users can only specify sane values.

+++ pkg/virt-launcher/virtwrap/live-migration-source.go
@@ -960,7 +960,7 @@ func (m *migrationMonitor) handleStallDetection(dom cli.VirDomain, stats *libvir
 	sd := m.stallDetector
 
 	if !sd.initialMaxDowntimeSet {
-		initialMaxDowntime := m.options.MaxDowntime
+		initialMaxDowntime := uint64(m.options.MaxDowntime)
 		if initialMaxDowntime > migrationutils.QEMUDefaultTargetDowntimeMS {
 			initialMaxDowntime = migrationutils.QEMUDefaultTargetDowntimeMS
 		}
